2024.08a

+++ Rust/test_code/src/main.rs
@@ -1,9 +1,9 @@
 fn main() {
-    use std::f32::consts::PI as PI;
-    const PI2: f32 = PI*2.0;
+    use std::f64::consts::PI as PI;
+    const PI2: f64 = PI*2.0;
 
-    let deg: f32 = 30.0;
-    let rad: f32 = (deg/360.0) * PI2;
-    let y: f32 = rad.cos();
+    let deg: f64 = 30.0;
+    let rad: f64 = (deg/360.0) * PI2;
+    let y: f64 = rad.cos();
     println!("{rad}, {y}");
 }
